requirements.txt
pandas
numpy
opencv-python
faiss-cpu
scikit-learn
tensorflow
keras